2021-10-15 sync

+++ incl/relationships/unblockGJUser.php
@@ -3,20 +3,15 @@ chdir(dirname(__FILE__));
 include "../lib/connection.php";
 require_once "../lib/GJPCheck.php";
 require_once "../lib/exploitPatch.php";
-$ep = new exploitPatch();
-$accountID = $ep->remove($_POST["accountID"]); //249 - maincra
-$gjp = $ep->remove($_POST["gjp"]);
-$targetAccountID = $ep->remove($_POST["targetAccountID"]); //250 - tomasek
-// REMOVING FOR USER 1
+
+if(empty($_POST['targetAccountID']))
+	exit("-1");
+
+$accountID = GJPCheck::getAccountIDOrDie();
+$targetAccountID = ExploitPatch::remove($_POST["targetAccountID"]);
+
 $query = "DELETE FROM blocks WHERE person1 = :accountID AND person2 = :targetAccountID";
 $query = $db->prepare($query);
-//EXECUTING THE QUERIES
-$GJPCheck = new GJPCheck();
-$gjpresult = $GJPCheck->check($gjp,$accountID);
-if($gjpresult == 1){
-	$query->execute([':accountID' => $accountID, ':targetAccountID' => $targetAccountID]);
-	echo "1";
-}else{
-	echo "-1";
-}
-?>
+$query->execute([':accountID' => $accountID, ':targetAccountID' => $targetAccountID]);
+
+echo "1";
